+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 #### Fixed
 - Refresh Token is not persisted when refresh_token grant is not issuing new Refresh Token [SDKS-649]
+- org.forgerock.android.auth.FRUser.getAccessToken() clean up tokens in the following conditions: [SDKS-701]
+-- When Refresh Token Grant Types is used, Server returns invalid_grant (Refresh Token expired), and failed to acquire OAuth2 Tokens with Session Token
+-- When Refresh Token Grant Types is not used, and failed to acquire OAuth2 Tokens with Session Token
 
 ## [2.1.0]